fixing on roles and permission

+++ app/Http/Controllers/Backend/UserController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers\Backend;
 
 use App\Http\Controllers\Controller;
+use App\Models\RoleHasPermission;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
@@ -12,17 +14,34 @@ class UserController extends Controller
     public function index()
     {
 
+        $permissionRole = RoleHasPermission::getPermissions('users', Auth::user()->role_id);
+        if (empty($permissionRole)) {
+            abort(404);
+        }
+
+        $PermissionAdd = RoleHasPermission::getPermissions('add-user', Auth::user()->role_id);
+        $PermissionEdit = RoleHasPermission::getPermissions('edit-user', Auth::user()->role_id);
+        $PermissionDelete = RoleHasPermission::getPermissions('delete-user', Auth::user()->role_id);
+
         $userRecord = User::getRecord();
         $roles = User::getRole();
 
         return view('backend.users.index', [
             'userRecord' => $userRecord,
-            'roles' => $roles
+            'roles' => $roles,
+            'PermissionAdd' => $PermissionAdd,
+            'PermissionEdit' => $PermissionEdit,
+            'PermissionDelete' => $PermissionDelete
         ]);
     }
 
     public function addUser()
     {
+        $permissionRole = RoleHasPermission::getPermissions('add-user', Auth::user()->role_id);
+        if (empty($permissionRole)) {
+            abort(404);
+        }
+
         return view('backend.users.add-user', [
             'roles' => User::getRole()
         ]);
@@ -35,6 +54,11 @@ class UserController extends Controller
         //     'password' => 'required|min:6',
         //     'role_id' => 'required|exists:roles,id', // Validate role_id
         // ]);
+
+        $permissionRole = RoleHasPermission::getPermissions('add-user', Auth::user()->role_id);
+        if (empty($permissionRole)) {
+            abort(404);
+        }
 
         $user = new User();
         $user->name = $request->name;
@@ -50,6 +74,11 @@ class UserController extends Controller
 
     public function editUser($id)
     {
+        $permissionRole = RoleHasPermission::getPermissions('edit-user', Auth::user()->role_id);
+        if (empty($permissionRole)) {
+            abort(404);
+        }
+
         return view('backend.users.edit-user', [
             'getRecord' => User::getSingle($id),
             'roles' => User::getRole()
@@ -71,6 +100,11 @@ class UserController extends Controller
 
     public function deleteUser($id)
     {
+        $permissionRole = RoleHasPermission::getPermissions('delete-user', Auth::user()->role_id);
+        if (empty($permissionRole)) {
+            abort(404);
+        }
+
         User::destroy($id);
         return redirect()->route('show.users')->with('success', 'Deleted Successfully');
     }
